Simplify query generation
Instead of manually enumerating all possible combinations of networks
and runtimes, creating queries for them, and then checking if they are
enabled, we can just compile the list of enabled combinations, and
generated queries just for them.

Doing this simplifies our code. (And now we can add more networks
and runtimes without manually having to touch this code.)

Normally, calling React hooks from callbacks and other conditional code
is not a good idea, but in this case, we can be sure that the number of
enabled combinations will never change during runtime
(since it's hard-coded in route-utils.ts), su we can just ignore
the TS warning about abusing the rules of hooks.