Add missing wrappers for API calls
Since we need to be able to override the URL for testnet,
and this functionality is currently only implemented in our manual
API wrappers, we need manual wrappers for all API functions we want
to use.

We already had them for most, but a few were missing, and thus
those calls always went to mainnet, instead of testnet.

This change fixes that.