Eliminate some useless token loading requests
When we are looking at account details, and the account is
a token contract, then we also need to load the token data
to get the name of the token, which we want to display.

Because of this, whenever looking at account details,
we were always requesting token data at the same time.

This commit restricts sending this request to cases
when we already know that we are looking at contract.