Support limiting scope to one network or layer
It's now possible to specify a given network or layer
in the .env file, and when given, the scope of the whole
explorer will be limited to that specific network,
or even a specific layer within that network.
No data external to the required scope will be shown.
The network (and the layer) will not be part of the URLs.

You can specify network, or layer, or both.