Treat mailto: protocol as special
Remove "mailto:" from the list of valid protocols,
because it can be unsafe to click on mailto links.

Supporting mailto: links can enable malicious URLs like
javascript: and mailto:?attach=.

We are trying to prevent that.

Instead, provide a special "emailAccepted" when we
know that the link is coming from a safe source,
like our code.