Tighter integration between text matching/trimming/highlighting
When we trim a text around a pattern match, and later we try
to highlight the match, pass the information along
from the trimming phase to the highlighting phase.

This avoids a round of unnecessary text matching, and fixes the corner case
when the trimmed part is actually shorter than the pattern.