Refactor LayerPicker
Include the name of the network for all layer menu items,
therefore eliminating the dependency of the layer menu
on the network menu.

The goal is that in the future, we can show the layer menu items
in other arrangements, not strictly following the separation
of the different networks.

Also, clean up internal architecture a naming a little bit.

There should be no functionality change.