Hide the disabled Pontus-X option from the layer selector on mainnet
Currently, the Pontus-X layer is enabled and visible
in the layer selector, because the Pontus-X Testnet runtime
is running on the Oasis Testnet, and we want to expose it
to the public.

However, we don't have a Pontus-X mainnet, and there are
no immediate plans to have one, so in the layer selector
for the mainnet, the Pontus-X option is disabled; it says
"coming soon".

I think it is not useful to have this disabled option here
(on the mainnet), but until now, we couldn't hide it,
because we wanted to see this layer on testnet.

However, we have recently gained the ability to show/hide
different layers (in the layer selector) per different
network, so now we can hide it from the mainnet, while
keeping it from the testnet. This commit does that.

x