When auto-redirecting from landing page, preserve scope
There are some network / layer configurations that
(if configured in the .env) result in not showing
the graph on the landing page, but instead immediately
redirecting to the dashboard of a Runtime.

This change makes it so that if we change to a layer
using the layer selector, that layer is preserved,
and next time we redirect from the landing page,
we go to this layer.

(Hint for testing: uncomment REACT_APP_FIXED_LAYER in .env)