Make adaptive trimmer more resilient
Up to now, the adaptive trimmer needed to be inserted in a DOM element
with a fixed width, and overflow-X set to hidden.
(This was required so that it can find the best size to fill it, detecting
overflow.)

This was fragile, because it relied on all parent components being configured
in a particular way.

This changes makes it so that we can detect overflow recursively
on any of the parent elements, which means that there are no
restrictions on parent configuration.

List of technical changes:
- New method for detecting overflow (on all parent elements)
- Check overflow status on short string as baseline,
  and then compare the situation with a longer string
- Introduce debug logging (on demand)
- Introduce a context-sensitive minimum length
- Support shortener functions that return a different length than requested
- Make sure to never elongate instead of shortening
- Reduce jumpy behavior (by hiding content during size discovery)
- Move the logic to a separate hook
- Introduce singleton controller, injected as context
- Set minimum lengths where required
- Add more extensive debug output
- Use more descriptive names for debugging
- Solve some wrapping issues related to highlighting