Separate searchQuery from search tokens
Until now, we have assumed that what we are searching
for is always identical to the search query input by the user.

As preparation for multi-token search, here we introduce
a new data structure for the parsed search query.

Both the actual search and the highlighting should be done
by looking at this data, instead of the original searchQuery.

This data structure will also hold any error messages
resulting from parsing the searchQuery.