Differentiate between different text searches
Up to now, we assumed all search works the same.
In order to gradually introduce multi-token search
for the various object types, this assumption is now relaxed.

From now on, whenever we want to interpret a searchQuery,
we have to specify what kind of search this is.
(I.e. searching for account names, or token names, etc.)