Fix testnet 404 text colors
Background: earlier, we had a 404 dialog with a bg that
was mostly transparent, so it was shown in changing dark and light
colors, depending on the current theme.

However, during a recent re-design, this concept has been dropped;
the bg is no longer transparent, it's always blue.

Therefore, the text color now always must be white.
However, this bit was missed during implementation, because
it looked OK on mainnet.

This PR fixes that: sets the text color to white,
independently of the current there.