Improve caching of token info
We are caching loaded token info. Cache is indexed by token address.
However, we can now reference token using both their oasis and eth addresses.

This change makes sure we are normalizing all addresses to oasis addresses,
so that we don't try to load the same info twice,
using different representations of the same address.

This corner case is triggered when we are on a screen when both versions are
on display, and we want to check for token names for both of them.

In these cases, up to this point, we were making some useless request to Nexus.
This change gets rid of that.